Move pkg/operation/seed/utils to pkg/utils/gardener
