+++ pkg/gardenlet/controller/seed/seed_control.go
@@ -34,7 +34,7 @@ import (
 	"github.com/gardener/gardener/pkg/operation/common"
 	"github.com/gardener/gardener/pkg/operation/garden"
 	seedpkg "github.com/gardener/gardener/pkg/operation/seed"
-	seedutils "github.com/gardener/gardener/pkg/operation/seed/utils"
+	gardenerutils "github.com/gardener/gardener/pkg/utils/gardener"
 	"github.com/gardener/gardener/pkg/utils/imagevector"
 	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
 
@@ -181,7 +181,7 @@ func (c *defaultControl) ReconcileSeed(obj *gardencorev1beta1.Seed, key string) 
 
 	seedNamespace := &corev1.Namespace{
 		ObjectMeta: metav1.ObjectMeta{
-			Name: seedutils.ComputeGardenNamespace(obj.Name),
+			Name: gardenerutils.ComputeGardenNamespace(obj.Name),
 		},
 	}
 
@@ -360,7 +360,7 @@ func (c *defaultControl) ReconcileSeed(obj *gardencorev1beta1.Seed, key string) 
 		ctx,
 		gardenClient.Client(),
 		c.seedLister,
-		seedutils.ComputeGardenNamespace(seed.Name),
+		gardenerutils.ComputeGardenNamespace(seed.Name),
 	)
 	if err != nil {
 		conditionSeedBootstrapped = gardencorev1beta1helper.UpdatedCondition(conditionSeedBootstrapped, gardencorev1beta1.ConditionFalse, "GardenSecretsError", err.Error())
